+++ packages/independent/assert/CHANGELOG.md
@@ -1,3 +1,44 @@
+# 2.11.0
+
+### add `assert.startsWith`
+
+Add a new helper method `assert.startsWith`. It is useful to perform assertion only on the beginning of a string and allow the rest to be anything. Very useful for error stack where the beginning is know but the rest of error stack is dependent on the filesystem an other things like node modules.
+
+`````md
+<!-- @jsenv/assert 2.10.0 -->
+
+```js
+const stack = `Error: message
+  at file.js:10:1
+  at node_modules/foo/foo.js:125:10`;
+const expected = `Error: message 
+  at files.js:10:1`;
+const actual = stack.slice(0, expected.length);
+assert({ actual, expected });
+```
+
+<!-- @jsenv/assert 2.11.0 -->
+
+```js
+const actual = `Error: message
+  at file.js:10:1
+  at node_modules/foo/foo.js:125:10`;
+const expected = assert.startsWith(`Error: message
+  at files.js:10:1`);
+assert({ actual, expected });
+```
+
+It was also possible to use `matchesRegExp` to obtain `startsWith` but again not very handy to write:
+
+```js
+const actual = `Error: message
+  at file.js:10:1
+  at node_modules/foo/foo.js:125:10`;
+const expected = /Error\: message\\n  at files.js\:10\:1.*+/;
+const actual = stack.slice(0, expected.length);
+assert({ actual, expected });
+```
+
 # 2.10.0
 
 #### Improve message for strings
@@ -43,6 +84,7 @@ my name is Damien"
 actual[18]#L2C12
 ```
 ````
+`````
 
 #### Limit message length for strings